+++ packages/capabilities/index.ts
@@ -1,3 +1,4 @@
 export * from './core/index.ts'
 export * from './market/index.ts'
+export * from './news/index.ts'
 export * from './providers/index.ts'